requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
crispy-tailwind==1.0.3
cryptography==43.0.3
Django==5.1.3
django-allauth==65.2.0
django-crispy-forms==2.3
django-request==1.6.3
idna==3.10
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.16.0
sqlparse==0.5.2
tzdata==2024.2
urllib3==2.2.3
